Refactoring of buildArgs
